+++ src/containers/Social/Layout.tsx
@@ -1,25 +1,29 @@
-import { useSelector } from "react-redux";
+import { useSelector } from 'react-redux';
 
-import LeftBar from "./sideBar/LeftBar";
-import RightBar from "./sideBar/RightBar";
+import LeftBar from './LeftBar';
+import RightBar from './RightBar';
 
-import type { RootState } from "@/common/redux/store";
+import type { RootState } from '@/common/redux/store';
 
-const Layout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
+interface PropsType {
+  children: React.ReactNode;
+}
+
+function Layout({ children }: PropsType) {
   const { userId } = useSelector((state: RootState) => state.userInfo);
 
   if (!userId) {
     return null;
   }
   return (
-    <section className="outer px-8">
-      <div className="inner relative flex justify-center">
+    <section className='outer px-8'>
+      <div className='inner relative flex justify-center'>
         <LeftBar />
         {children}
         <RightBar />
       </div>
     </section>
   );
-};
+}
 
 export default Layout;
